Move `Util.kt` to `query.compose.util` package
As the amount of Util.kt code has grown, we decided to create a dedicated package space for it.
