+++ internal/playground/src/commonMain/kotlin/soil/playground/query/compose/RememberExampleSubscription.kt
@@ -5,8 +5,8 @@ import androidx.compose.runtime.remember
 import soil.playground.query.key.ExampleSubscriptionKey
 import soil.query.annotation.ExperimentalSoilQueryApi
 import soil.query.compose.SubscriptionObject
-import soil.query.compose.auto
 import soil.query.compose.rememberSubscription
+import soil.query.compose.util.auto
 import soil.query.core.Namespace
 
 @OptIn(ExperimentalSoilQueryApi::class)
